[CHANGE] file startup_stm32f103xx.s and modify TIM6_DAC_IRQHanler, TIM6_DAC_IRQn

+++ Microcontroller/STM32F1xx/Driver/source/cortex_m3.c
@@ -83,7 +83,7 @@ void NVIC_EnableIRQ(IRQn_Type IRQn)
 {
   if (IRQn >= 0)
   {
-    NVIC.NVIC_ISER[IRQn >> 5] = (1 << (IRQn & 0x1F)); //Set bit to enable IRQ
+    NVIC.NVIC_ISER[IRQn >> 5] = (1 << (((uint32_t)IRQn) & 0x1F)); //Set bit to enable IRQ
   }
 }
